move components to separate files
