+++ django_q/management/commands/qmonitor.py
@@ -1,7 +1,7 @@
 from django.core.management.base import BaseCommand
 from django.utils.translation import gettext as _
 
-from django_q.monitor import monitor
+from django_q.monitor_terminal import monitor
 
 
 class Command(BaseCommand):
