+++ django_q/signals.py
@@ -31,6 +31,7 @@ def call_hook(sender, instance, **kwargs):
                 % {"hook": instance.hook, "name": instance.name, "error": str(e)}
             )
 
+
 # args: proc_name
 post_spawn = Signal()
